Extract class ModuleSymbol from symbol.{h,cpp} in to ModuleSymbol.{h,cpp}
