+++ compiler/AST/flags.cpp
@@ -1,15 +1,15 @@
 /*
  * Copyright 2004-2017 Cray Inc.
  * Other additional copyright holders may be indicated within.
- * 
+ *
  * The entirety of this work is licensed under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
- * 
+ *
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -73,31 +73,41 @@ void writeFlags(FILE* fp, Symbol* sym) {
 
 
 // these affect what viewFlags() prints
-bool viewFlagsShort = true;
-bool viewFlagsPragma = false;
-bool viewFlagsName  = false;
+bool viewFlagsShort   = true;
+bool viewFlagsPragma  = false;
+bool viewFlagsName    = false;
 bool viewFlagsComment = false;
-bool viewFlagsExtras = true;
+bool viewFlagsExtras  = true;
 
-void
-viewFlags(BaseAST* ast) {
-  if (!viewFlagsShort && !viewFlagsName && !viewFlagsComment)
+void viewFlags(BaseAST* ast) {
+  if (!viewFlagsShort && !viewFlagsName && !viewFlagsComment) {
     viewFlagsName = true;
+  }
+
   if (Symbol* sym = toSymbol(ast)) {
     for (int flagNum = FLAG_FIRST; flagNum <= FLAG_LAST; flagNum++) {
       if (sym->flags[flagNum]) {
-        if (viewFlagsName)
+        if (viewFlagsName) {
           printf("%s ", flagNames[flagNum]);
-        if (viewFlagsPragma)
+        }
+
+        if (viewFlagsPragma) {
           printf("%s", flagPragma[flagNum] ? "ypr " : "npr ");
-        if (viewFlagsShort)
+        }
+
+        if (viewFlagsShort) {
           printf("\"%s\" ", flagShortNames[flagNum]);
-        if (viewFlagsComment)
+        }
+
+        if (viewFlagsComment) {
           printf("// %s",
                  *flagComments[flagNum] ? flagComments[flagNum] : "ncm");
+        }
+
         printf("\n");
       }
     }
+
     if (viewFlagsExtras) {
       if (VarSymbol* vs = toVarSymbol(sym)) {
         if (vs->immediate) {
@@ -105,24 +115,33 @@ viewFlags(BaseAST* ast) {
           fprint_imm(stdout, *toVarSymbol(sym)->immediate, true);
           printf("\n");
         }
+
       } else if (ArgSymbol* as = toArgSymbol(sym)) {
         printf("%s arg\n", as->intentDescrString());
+
       } else if (toTypeSymbol(sym)) {
         printf("a TypeSymbol\n");
+
       } else if (FnSymbol* fs = toFnSymbol(sym)) {
         printf("fn %s(%d args) %s\n",
                fs->_this ? intentDescrString(fs->thisTag) : "",
-               fs->numFormals(), retTagDescrString(fs->retTag));
+               fs->numFormals(),
+               retTagDescrString(fs->retTag));
+
       } else if (toEnumSymbol(sym)) {
         printf("an EnumSymbol\n");
+
       } else if (ModuleSymbol* ms = toModuleSymbol(sym)) {
-        printf("module %s\n", modTagDescrString(ms->modTag));
+        printf("module %s\n", ModuleSymbol::modTagToString(ms->modTag));
+
       } else if (toLabelSymbol(sym)) {
         printf("a LabelSymbol\n");
+
       } else {
         printf("unknown symbol kind\n");
       }
     }
+
   } else {
     printf("[%d]: not a Symbol, has no flags\n", ast->id);
   }
